Ensure that doom's customizations to straight.el take effect
`nix-straight.el` used to require `straight.el` on its own before `doom`
itself was loaded. This had an unfortunate side-effect that
`straight-fix-org` was set to nil after it has been loaded so
`org-version` and `org-git-version` functions were already overloaded.
This was no problem for `org-mode` itself as it has its own overload of
the above mentioned functions, but other libraries such as `elfeed` may
also use these which causes problems.

`nix-straight.el` is updated to a version where it does not preload
`straight` and an advice is provided to `doom` so it loads our version
of `straight` when it supposed to.

References: https://github.com/vlaci/nix-doom-emacs/issues/5